+++ src/pydantic_market_data/interfaces.py
@@ -7,9 +7,9 @@ from .models import (
     History,
     HistoryPeriod,
     Price,
+    Security,
     SecurityCriteria,
     Symbol,
-    Ticker,
 )
 
 
@@ -18,30 +18,30 @@ class DataSource(Protocol):
     Interface for a financial data source.
     """
 
-    def search(self, query: str) -> list[Symbol]:
+    def search(self, query: str) -> list[Security]:
         """
         Search for security by ISIN, symbol, or name
         """
         ...
 
-    def resolve(self, criteria: SecurityCriteria) -> Symbol | None:
+    def resolve(self, criteria: SecurityCriteria) -> Security | None:
         """
         Resolve security based on provided criteria
         """
         ...
 
-    def history(self, ticker: Ticker.Input, period: HistoryPeriod = HistoryPeriod.MO1) -> History:
+    def history(self, symbol: Symbol.Input, period: HistoryPeriod = HistoryPeriod.MO1) -> History:
         """
-        Fetch historical data for a ticker
+        Fetch historical data for a symbol
         """
         ...
 
-    def get_price(self, ticker: Ticker.Input, date: date | None = None) -> Price | None:
-        """Fetch the price for a ticker (current or historical)"""
+    def get_price(self, symbol: Symbol.Input, date: date | None = None) -> Price | None:
+        """Fetch the price for a symbol (current or historical)"""
         ...
 
-    def validate(self, ticker: Ticker.Input, target_date: date, target_price: Price.Input) -> bool:
+    def validate(self, symbol: Symbol.Input, target_date: date, target_price: Price.Input) -> bool:
         """
-        Validates if the ticker traded near the target price on the target date.
+        Validates if the symbol traded near the target price on the target date.
         """
         ...